move npm files
